+++ src/main/groovy/com/moowork/gradle/node/npm/NpmSetupTask.groovy
@@ -1,10 +1,12 @@
 package com.moowork.gradle.node.npm
 
 import com.moowork.gradle.node.NodeExtension
+import com.moowork.gradle.node.NodePlugin
 import com.moowork.gradle.node.task.SetupTask
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Internal
+import org.gradle.api.tasks.Nested
 import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.TaskAction
 import org.gradle.process.ExecResult
@@ -29,7 +31,7 @@ class NpmSetupTask
     {
         dependsOn( SetupTask.NAME )
 
-        this.group = 'Node'
+        this.group = NodePlugin.NODE_GROUP
         this.description = 'Setup a specific version of npm to be used by the build.'
         this.enabled = false
 
@@ -37,7 +39,7 @@ class NpmSetupTask
     }
 
     @Input
-    Set<String> getInput()
+    Set<Object> getInput()
     {
         def set = new HashSet<>()
         set.add( getConfig().download )
@@ -76,12 +78,12 @@ class NpmSetupTask
         return getConfig().variant
     }
 
+    @Input
     List<?> getArgs()
     {
         return this.args
     }
 
-    @Internal
     void setArgs( final Iterable<?> value )
     {
         this.args = value.toList()
@@ -95,6 +97,12 @@ class NpmSetupTask
     void setExecOverrides( final Closure closure )
     {
         this.runner.execOverrides = closure
+    }
+
+    @Nested
+    NpmExecRunner getRunner()
+    {
+        return runner
     }
 
     @TaskAction