+++ src/main/groovy/com/moowork/gradle/node/task/SetupTask.groovy
@@ -1,10 +1,10 @@
 package com.moowork.gradle.node.task
 
 import com.moowork.gradle.node.NodeExtension
+import com.moowork.gradle.node.NodePlugin
 import com.moowork.gradle.node.util.BackwardsCompat
 import com.moowork.gradle.node.variant.Variant
 import org.gradle.api.DefaultTask
-import org.gradle.api.artifacts.repositories.ArtifactRepository
 import org.gradle.api.artifacts.repositories.IvyArtifactRepository
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Internal
@@ -24,13 +24,10 @@ class SetupTask
 
     protected Variant variant
 
-    private IvyArtifactRepository repo
-
-    private List<ArtifactRepository> allRepos;
-
     SetupTask()
     {
-        this.group = 'Node'
+
+        this.group = NodePlugin.NODE_GROUP
         this.description = 'Download and install a local node/npm version.'
         this.enabled = false
     }
@@ -69,7 +66,7 @@ class SetupTask
     void exec()
     {
         configureIfNeeded()
-        addRepository()
+        addRepositoryIfNeeded()
 
         if ( this.variant.exeDependency )
         {
@@ -79,7 +76,6 @@ class SetupTask
         deleteExistingNode()
         unpackNodeArchive()
         setExecutableFlag()
-        restoreRepositories()
     }
 
     private void copyNodeExe()
@@ -169,18 +165,25 @@ class SetupTask
         return conf.resolve().iterator().next();
     }
 
-    private void addRepository()
-    {
-        this.allRepos = new ArrayList<>()
-        this.allRepos.addAll( this.project.repositories )
-        this.project.repositories.clear()
+    private void addRepositoryIfNeeded() {
+        if ( this.config.distBaseUrl != null ) {
+            addRepository this.config.distBaseUrl
+        }
+    }
 
-        def distUrl = this.config.distBaseUrl
-        this.repo = this.project.repositories.ivy {
+    private void addRepository( String distUrl ) {
+        this.project.repositories.ivy {
             url distUrl
-            layout 'pattern', {
-                artifact 'v[revision]/[artifact](-v[revision]-[classifier]).[ext]'
-                ivy 'v[revision]/ivy.xml'
+            if (BackwardsCompat.usePatternLayout()) {
+                patternLayout {
+                    artifact 'v[revision]/[artifact](-v[revision]-[classifier]).[ext]'
+                    ivy 'v[revision]/ivy.xml'
+                }
+            } else {
+                layout 'pattern', {
+                    artifact 'v[revision]/[artifact](-v[revision]-[classifier]).[ext]'
+                    ivy 'v[revision]/ivy.xml'
+                }
             }
             if (BackwardsCompat.useMetadataSourcesRepository()) {
                 metadataSources {
@@ -188,11 +191,5 @@ class SetupTask
                 }
             }
         }
-    }
-
-    private void restoreRepositories()
-    {
-        this.project.repositories.clear();
-        this.project.repositories.addAll( this.allRepos );
     }
 }