+++ src/test/groovy/com/moowork/gradle/node/npm/NpxTask_integTest.groovy
@@ -1,0 +1,167 @@
+package com.moowork.gradle.node.npm
+
+import com.moowork.gradle.AbstractIntegTest
+import org.gradle.testkit.runner.TaskOutcome
+import org.junit.Rule
+import org.junit.contrib.java.lang.system.EnvironmentVariables
+
+class NpxTask_integTest
+        extends AbstractIntegTest {
+    @Rule
+    EnvironmentVariables environmentVariables = new EnvironmentVariables()
+
+    def 'execute npx command with no package.json file'() {
+        given:
+        writeBuild('''
+            plugins {
+                id 'com.github.node-gradle.node'
+            }
+
+            node {
+                version = "10.14.0"
+                npmVersion = "6.4.1"
+                download = true
+                workDir = file('build/node')
+            }
+            
+            task camelCase(type: NpxTask) {
+                command = 'chcase-cli'
+                args = ['--help']
+            }
+        ''')
+
+        when:
+        def result = build(":camelCase")
+
+        then:
+        result.task(":nodeSetup").outcome == TaskOutcome.SUCCESS
+        result.task(":npmSetup").outcome == TaskOutcome.SUCCESS
+        result.task(":camelCase").outcome == TaskOutcome.SUCCESS
+        result.output.contains("--case, -C  Which case to convert to")
+    }
+
+    def 'execute npx command with a package.json file and check inputs up-to-date detection'() {
+        given:
+        copyResources('fixtures/npx/', '')
+        copyResources('fixtures/javascript-project/', '')
+
+        when:
+        def result1 = build(":test")
+
+        then:
+        result1.task(":nodeSetup").outcome == TaskOutcome.SUCCESS
+        result1.task(":npmSetup").outcome == TaskOutcome.SUCCESS
+        result1.task(":npmInstall").outcome == TaskOutcome.SUCCESS
+        result1.task(":lint").outcome == TaskOutcome.SUCCESS
+        result1.task(":test").outcome == TaskOutcome.SUCCESS
+        result1.output.contains("3 problems (0 errors, 3 warnings)")
+        result1.output.contains("1 passing")
+
+        when:
+        def result2 = build(":test")
+
+        then:
+        result2.task(":nodeSetup").outcome == TaskOutcome.UP_TO_DATE
+        result2.task(":npmSetup").outcome == TaskOutcome.UP_TO_DATE
+        result2.task(":npmInstall").outcome == TaskOutcome.SUCCESS
+        result2.task(":lint").outcome == TaskOutcome.UP_TO_DATE
+        result2.task(":test").outcome == TaskOutcome.UP_TO_DATE
+
+        when:
+        def result3 = build(":test", "-DchangeInputs=true")
+
+        then:
+        result3.task(":nodeSetup").outcome == TaskOutcome.UP_TO_DATE
+        result3.task(":npmSetup").outcome == TaskOutcome.UP_TO_DATE
+        result3.task(":npmInstall").outcome == TaskOutcome.UP_TO_DATE
+        result3.task(":lint").outcome == TaskOutcome.SUCCESS
+        result3.task(":test").outcome == TaskOutcome.SUCCESS
+    }
+
+    def 'execute npx command with custom execution configuration and check up-to-date-detection'() {
+        given:
+        copyResources('fixtures/npx-env/', '')
+        copyResources('fixtures/env/', '')
+
+        when:
+        def result1 = build(":env")
+
+        then:
+        result1.task(":nodeSetup").outcome == TaskOutcome.SUCCESS
+        result1.task(":npmSetup").outcome == TaskOutcome.SUCCESS
+        result1.task(":npmInstall").outcome == TaskOutcome.SUCCESS
+        result1.task(":env").outcome == TaskOutcome.SUCCESS
+        result1.output.contains("PATH=")
+
+        when:
+        def result2 = build(":env", "-DcustomEnv=true")
+
+        then:
+        result2.task(":nodeSetup").outcome == TaskOutcome.UP_TO_DATE
+        result2.task(":npmSetup").outcome == TaskOutcome.UP_TO_DATE
+        result2.task(":npmInstall").outcome == TaskOutcome.SUCCESS
+        result2.task(":env").outcome == TaskOutcome.SUCCESS
+        result2.output.contains("CUSTOM=custom value")
+
+        when:
+        environmentVariables.set("NEW_ENV_VARIABLE", "Let's make the whole environment change")
+        def result3 = build(":env", "-DcustomEnv=true")
+
+        then:
+        result3.task(":nodeSetup").outcome == TaskOutcome.UP_TO_DATE
+        result3.task(":npmSetup").outcome == TaskOutcome.UP_TO_DATE
+        result3.task(":npmInstall").outcome == TaskOutcome.UP_TO_DATE
+        result3.task(":env").outcome == TaskOutcome.UP_TO_DATE
+
+        when:
+        def result4 = build(":env", "-DignoreExitValue=true", "-DnotExistingCommand=true")
+
+        then:
+        result4.task(":nodeSetup").outcome == TaskOutcome.UP_TO_DATE
+        result4.task(":npmSetup").outcome == TaskOutcome.UP_TO_DATE
+        result4.task(":npmInstall").outcome == TaskOutcome.UP_TO_DATE
+        result4.task(":env").outcome == TaskOutcome.SUCCESS
+        result4.output.contains("E404")
+
+        when:
+        def result5 = buildAndFail(":env", "-DnotExistingCommand=true")
+
+        then:
+        result5.task(":nodeSetup").outcome == TaskOutcome.UP_TO_DATE
+        result5.task(":npmSetup").outcome == TaskOutcome.UP_TO_DATE
+        result5.task(":npmInstall").outcome == TaskOutcome.UP_TO_DATE
+        result5.task(":env").outcome == TaskOutcome.FAILED
+        result5.output.contains("E404")
+
+        when:
+        def result6 = build(":pwd")
+
+        then:
+        result6.task(":nodeSetup").outcome == TaskOutcome.UP_TO_DATE
+        result6.task(":npmSetup").outcome == TaskOutcome.UP_TO_DATE
+        result6.task(":npmInstall").outcome == TaskOutcome.UP_TO_DATE
+        result6.task(":pwd").outcome == TaskOutcome.SUCCESS
+        result6.output.contains("Working directory is '${projectDir}'")
+
+        when:
+        def result7 = build(":pwd", "-DcustomWorkingDir=true")
+
+        then:
+        result7.task(":nodeSetup").outcome == TaskOutcome.UP_TO_DATE
+        result7.task(":npmSetup").outcome == TaskOutcome.UP_TO_DATE
+        result7.task(":npmInstall").outcome == TaskOutcome.UP_TO_DATE
+        result7.task(":pwd").outcome == TaskOutcome.UP_TO_DATE
+
+        when:
+        def result8 = build(":pwd", "-DcustomWorkingDir=true", "--rerun-tasks")
+
+        then:
+        result8.task(":nodeSetup").outcome == TaskOutcome.SUCCESS
+        result8.task(":npmSetup").outcome == TaskOutcome.SUCCESS
+        result8.task(":npmInstall").outcome == TaskOutcome.SUCCESS
+        result8.task(":pwd").outcome == TaskOutcome.SUCCESS
+        def expectedWorkingDirectory = "${projectDir}${File.separator}build${File.separator}customWorkingDirectory"
+        result8.output.contains("Working directory is '${expectedWorkingDirectory}'")
+        new File(expectedWorkingDirectory).isDirectory()
+    }
+}